Refactor process deletion controls and fix memory allocation status
