+++ src/components/body/simulator/cpu/CPUProcess.tsx
@@ -1,4 +1,4 @@
-import { Trash2 } from "lucide-react";
+import DeleteProcessButton from "../DeleteProcessButton";
 
 const statusStyles = {
   EXECUTING: "text-primary",
@@ -27,17 +27,6 @@ const ProcessStatus = ({ status }: { status: ProcessStatusValue }) => (
     )}
     {status.toLowerCase()}
   </span>
-);
-
-const DeleteProcessButton = ({ name, onDelete }: { name: string; onDelete: () => void }) => (
-  <button
-    type="button"
-    className="inline-flex h-7 w-7 items-center justify-center rounded-[3px] text-muted-foreground transition-colors hover:bg-destructive/10 hover:text-destructive focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-    aria-label={`Delete ${name}`}
-    onClick={onDelete}
-  >
-    <Trash2 className="h-3.5 w-3.5" strokeWidth={1.75} />
-  </button>
 );
 
 const CPUProcess = ({
